chore: divided component for smaller subcomponents
